fix dtype issues

+++ model.py
@@ -60,7 +60,7 @@ class Transformer(nn.Module):
             self.dtype = dtype
         self.w_embs = nn.Embedding(vocab_size, model_dim).to(self.dtype)
         self.blocks = nn.ModuleList([Block(model_dim, num_heads) for _ in range(num_layers)])
-        self.classifier = nn.Linear(model_dim, vocab_size, bias=False).to(self.dtype)
+        self.classifier = nn.Linear(model_dim, vocab_size, bias=False) # .to(self.dtype) this broke everything
 
     def forward(self, x, targets = None):
         x = self.w_embs(x)
